move helpers and fix invalid date info
