+++ react_app/src/components/CheckEmptiness/CheckEmptiness.module.scss
@@ -1,4 +1,4 @@
-@import "../helpers/_variables";
+@import "../../helpers/_variables";
 
 
 div {
