+++ react_app/src/components/Items/Items.module.scss
@@ -1,5 +1,5 @@
-@import "../helpers/_variables";
-@import "../helpers/helpers";
+@import "../../helpers/_variables";
+@import "../../helpers/helpers";
 
 
 .main-content {
