+++ react_app/src/components/Menu/ToggleMenu.module.scss
@@ -1,5 +1,5 @@
-@import "../helpers/_variables";
-@import "../helpers/helpers";
+@import "../../helpers/_variables";
+@import "../../helpers/helpers";
 
 
 .toggle-menu {
