+++ react_app/src/components/TaskForm/TaskForm.module.scss
@@ -1,5 +1,5 @@
-@import "../helpers/helpers";
-@import "../helpers/_variables";
+@import "../../helpers/helpers";
+@import "../../helpers/_variables";
 
 
 .form-container {
