+++ react_app/src/helpers/Interfaces.ts
