+++ react_app/src/helpers/_variables.scss
