+++ react_app/src/helpers/helpers.scss
